moved helper functions
